feat: fully functional k8s example!
As an added benefit, I dropped a bunch of non-trivial code,
that was trying to do what k8s already does.

Much easier that trying to replicate it just for the sake
of not relying on K8S.

Anybody can either run k3s on a linux/bsd box,
or docker-desktop on windows/macos.